requirements.txt
quart==0.20.0
line-bot-sdk==3.17.1
python-dotenv==1.1.0
discord.py==2.3.2
requests==2.32.3
discohook